// src/types/types.ts
export interface RegisterFormData {
    fullName: string
    email: string
    phone: string
    dni: string
    edad: string
    profesion: string
    church: string
    churchRole: string
    area: string
    district: string
    voucherCode: string
    voucherFile: File | null
    workshopDay1: string
    workshopDay2: string
}